fix: resolve startup errors for JWT, DB connection, and schema validation
- JwtTokenProvider: catch Exception instead of IllegalArgumentException
  so the UTF-8 fallback fires when JJWT throws DecodingException for
  non-Base64 secrets (e.g. plain dev strings with hyphens)
- application-dev.yml: use hrss_user/dev_app_password to match Docker
  MySQL MYSQL_USER setup; update MySQL port to 3307 (Docker host port);
  remove hyphens from JWT dev secret to avoid Base64 decode errors
- Attendance: change workHours from Double to BigDecimal to match
  DECIMAL(5,2) column type in Flyway migration (Double maps to FLOAT
  which caused Hibernate schema-validation failure)
- AttendanceDto: align workHours type to BigDecimal for consistency

+++ src/main/java/com/mod/hrss/dto/response/AttendanceDto.java
@@ -3,6 +3,7 @@ package com.mod.hrss.dto.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,5 +14,5 @@ public class AttendanceDto {
     private String employeeName;
     private LocalDateTime checkIn;
     private LocalDateTime checkOut;
-    private Double workHours;
+    private BigDecimal workHours;
 }
